pyproject.toml
[project]
name = "podcast-automation"
version = "1.3.0"
requires-python = ">=3.12"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    # AI/ML packages
    "openai>=1.0.0",
    "openai-whisper>=20231117",
    "whisperx>=3.1.6",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    # Audio processing
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "mutagen>=1.47.0",
    "pysubs2>=1.8.0",
    # Image processing
    "pillow>=10.2.0",
    # Cloud storage
    "dropbox>=12.0.2",
    # Social media APIs
    "google-api-python-client>=2.116.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "requests>=2.31.0",
    "tweepy>=4.14.0",
    # Reddit API
    "praw>=7.7.1",
    # Utilities
    "tqdm>=4.66.1",
    "pyyaml>=6.0.1",
    # Episode webpages
    "pygithub>=2.0.0",
    "yake>=0.4.8",
    "jinja2>=3.0.0",
    # Testing
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
